Созданы модули. Код разнесен по модулям.
